# Use reflect.Select for WatchSet