Instead of a set of goroutines doing select use reflect.Select instead as it's a bit faster and allocates less.

Benchmark before:
```
    goos: linux
    goarch: arm64
    pkg: github.com/cilium/statedb
    BenchmarkWatchSet_4
    BenchmarkWatchSet_4-6             633849              2027 ns/op             736 B/op         15 allocs/op
    BenchmarkWatchSet_16
    BenchmarkWatchSet_16-6            273488              4521 ns/op            1232 B/op         18 allocs/op
    BenchmarkWatchSet_128
    BenchmarkWatchSet_128-6            48825             24887 ns/op            5653 B/op         28 allocs/op
    BenchmarkWatchSet_1024
    BenchmarkWatchSet_1024-6            5077            262523 ns/op           55920 B/op        109 allocs/op
```

Benchmark after:
```
    goos: linux
    goarch: arm64
    pkg: github.com/cilium/statedb
    BenchmarkWatchSet_4
    BenchmarkWatchSet_4-6            1493523               918.2 ns/op           499 B/op          7 allocs/op
    BenchmarkWatchSet_16
    BenchmarkWatchSet_16-6            480783              2105 ns/op            1700 B/op          8 allocs/op
    BenchmarkWatchSet_128
    BenchmarkWatchSet_128-6            68299             19760 ns/op           13400 B/op          8 allocs/op
    BenchmarkWatchSet_1024
    BenchmarkWatchSet_1024-6            6812            174220 ns/op          105176 B/op          8 allocs/op
```